# doc/api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Recetas","description":"Operaciones relacionadas con recetas de yogurt"}],"paths":{"/api/recipes/{id}":{"get":{"tags":["Recetas"],"summary":"Obtener receta por ID","description":"Devuelve una receta específica a partir de su identificador único","operationId":"getRecipe","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Receta encontrada correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Recipe"}}}},"404":{"description":"No se encontró la receta con el ID proporcionado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Recipe"}}}}}},"put":{"tags":["Recetas"],"summary":"Actualizar receta","operationId":"updateRecipe","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecipeDto"}}},"required":true},"responses":{"200":{"description":"Receta actualizada correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Recipe"}}}},"404":{"description":"No se encontró la receta con el ID proporcionado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Recipe"}}}}}}},"/api/recipes":{"get":{"tags":["Recetas"],"summary":"Obtener todas las recetas","operationId":"getAllRecipes","responses":{"200":{"description":"Recetas encontradas correctamente","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Recipe"}}}}}}},"post":{"tags":["Recetas"],"summary":"Crear receta","operationId":"createRecipe","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecipeDto"}}},"required":true},"responses":{"201":{"description":"Receta creada correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Recipe"}}}},"400":{"description":"Datos de la receta inválidos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Recipe"}}}}}}},"/api/batches":{"get":{"tags":["yogurt-batch-controller"],"summary":"Obtener todos los lotes de yogurt","operationId":"getAllBatches","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["PREPARING","HEATING","COOLING","INOCULATING","INCUBATION","REFRIGERATION","COMPLETED","FAILED","REFRIGERATING","INCUBATING"]}}],"responses":{"200":{"description":"Lotes de yogurt obtenidos correctamente","content":{"*/*":{"schema":{"type":"object"}}}},"400":{"description":"Solicitud inválida","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["yogurt-batch-controller"],"summary":"Iniciar un nuevo lote de yogurt","operationId":"startNewBatch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StartBatchRequest"}}},"required":true},"responses":{"201":{"description":"Lote de yogurt creado correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/YogurtBatch"}}}},"400":{"description":"Solicitud inválida","content":{"*/*":{"schema":{"$ref":"#/components/schemas/YogurtBatch"}}}}}}},"/api/batches/{batchId}/refrigeration":{"post":{"tags":["yogurt-batch-controller"],"summary":"Iniciar el proceso de refrigeración para un lote de yogurt","operationId":"startRefrigeration","parameters":[{"name":"batchId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"201":{"description":"Proceso de refrigeración iniciado correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/YogurtBatch"}}}},"400":{"description":"Solicitud inválida","content":{"*/*":{"schema":{"$ref":"#/components/schemas/YogurtBatch"}}}}}}},"/api/batches/{batchId}/inoculating":{"post":{"tags":["yogurt-batch-controller"],"summary":"Iniciar el proceso de inoculación para un lote de yogurt","operationId":"startInoculating","parameters":[{"name":"batchId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"201":{"description":"Proceso de inoculación iniciado correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/YogurtBatch"}}}},"400":{"description":"Solicitud inválida","content":{"*/*":{"schema":{"$ref":"#/components/schemas/YogurtBatch"}}}}}}},"/api/batches/{batchId}/incubation":{"post":{"tags":["yogurt-batch-controller"],"summary":"Iniciar el proceso de incubación para un lote de yogurt","operationId":"startIncubation","parameters":[{"name":"batchId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"201":{"description":"Proceso de incubación iniciado correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/YogurtBatch"}}}},"400":{"description":"Solicitud inválida","content":{"*/*":{"schema":{"$ref":"#/components/schemas/YogurtBatch"}}}}}}},"/api/batches/{batchId}/heating":{"post":{"tags":["yogurt-batch-controller"],"summary":"Iniciar el proceso de calentamiento para un lote de yogurt","operationId":"startHeating","parameters":[{"name":"batchId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"201":{"description":"Proceso de calentamiento iniciado correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/YogurtBatch"}}}},"400":{"description":"Solicitud inválida","content":{"*/*":{"schema":{"$ref":"#/components/schemas/YogurtBatch"}}}}}}},"/api/batches/{batchId}/fail":{"post":{"tags":["yogurt-batch-controller"],"summary":"Marcar un lote de yogurt como fallido","operationId":"markAsFailed","parameters":[{"name":"batchId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FailRequest"}}},"required":true},"responses":{"200":{"description":"Lote de yogurt marcado como fallido correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/YogurtBatch"}}}},"400":{"description":"Solicitud inválida","content":{"*/*":{"schema":{"$ref":"#/components/schemas/YogurtBatch"}}}}}}},"/api/batches/{batchId}/complete":{"post":{"tags":["yogurt-batch-controller"],"summary":"Completar un lote de yogurt","operationId":"completeBatch","parameters":[{"name":"batchId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Lote de yogurt completado correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/YogurtBatch"}}}},"400":{"description":"Solicitud inválida","content":{"*/*":{"schema":{"$ref":"#/components/schemas/YogurtBatch"}}}}}}},"/api/recipes/{id}/deactivate":{"patch":{"tags":["Recetas"],"summary":"Desactivar receta","operationId":"deactivateRecipe","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Receta desactivada correctamente"},"404":{"description":"No se encontró la receta con el ID proporcionado"}}}},"/api/recipes/{id}/activate":{"patch":{"tags":["Recetas"],"summary":"Activar receta","operationId":"activateRecipe","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Receta activada correctamente"},"404":{"description":"No se encontró la receta con el ID proporcionado"}}}},"/api/recipes/search":{"get":{"tags":["Recetas"],"summary":"Buscar recetas","operationId":"searchRecipes","parameters":[{"name":"keyword","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Recetas encontradas correctamente","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Recipe"}}}}}}}},"/api/monitoring/dashboard":{"get":{"tags":["monitoring-controller"],"summary":"Obtener panel de control de monitoreo","operationId":"getDashboard","responses":{"200":{"description":"Panel de control obtenido correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Dashboard"}}}},"404":{"description":"No se encontraron lotes activos","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Dashboard"}}}}}}},"/api/monitoring/batches/{batchId}/temperature":{"get":{"tags":["monitoring-controller"],"summary":"Obtener resumen de temperatura para un lote de yogurt","operationId":"getBatchTemperatureSummary","parameters":[{"name":"batchId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Resumen de temperatura obtenido correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TemperatureSummary"}}}},"404":{"description":"No se encontró el lote de yogurt con el ID proporcionado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TemperatureSummary"}}}}}}},"/api/monitoring/batches/{batchId}/temperature-logs":{"get":{"tags":["monitoring-controller"],"summary":"Obtener registros de temperatura para un lote de yogurt","operationId":"getTemperatureLogs","parameters":[{"name":"batchId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"start","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"end","in":"query","required":false,"schema":{"type":"string","format":"date-time"}}],"responses":{"200":{"description":"Registros de temperatura obtenidos correctamente","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TemperatureLog"}}}}},"404":{"description":"No se encontró el lote de yogurt con el ID proporcionado","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TemperatureLog"}}}}}}}},"/api/monitoring/batches/active":{"get":{"tags":["monitoring-controller"],"summary":"Obtener lotes de yogurt activos","operationId":"getActiveBatches","responses":{"200":{"description":"Lotes encontrados correctamente","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/YogurtBatch"}}}}},"404":{"description":"No se encontraron lotes activos","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/YogurtBatch"}}}}}}}}},"components":{"schemas":{"IngredientDto":{"type":"object","properties":{"name":{"type":"string","description":"Nombre del ingrediente","example":"Azúcar"},"quantity":{"type":"number","format":"double","description":"Cantidad del ingrediente","example":100},"unit":{"type":"string","description":"Unidad de medida del ingrediente","example":"g"},"notes":{"type":"string","description":"Notas adicionales sobre el ingrediente","example":"Usar azúcar moreno para un sabor más intenso"},"optional":{"type":"boolean","description":"Indica si el ingrediente es opcional","example":false}}},"RecipeDto":{"type":"object","properties":{"name":{"type":"string","description":"ID de la receta","example":"1"},"description":{"type":"string","description":"Descripción de la receta","example":"Receta clásica de yogurt natural con un sabor suave y cremoso"},"defaultMilkVolume":{"type":"number","format":"double","description":"Volumen de leche recomendado para la receta","example":1.5},"defaultStarterAmount":{"type":"number","format":"double","description":"Cantidad de cultivo iniciador recomendado para la receta","example":2},"heatingTemperature":{"type":"number","format":"double","description":"Temperatura de calentamiento recomendada para la receta","example":85},"heatingDuration":{"type":"integer","format":"int32","description":"Duración del calentamiento recomendada para la receta","example":30},"inoculationTemperature":{"type":"number","format":"double","description":"Temperatura de inoculación recomendada para la receta","example":45},"incubationTemperature":{"type":"number","format":"double","description":"Temperatura de incubación recomendada para la receta","example":42},"minIncubationTime":{"type":"integer","format":"int32","description":"Tiempo mínimo de incubación recomendado para la receta en minutos","example":480},"maxIncubationTime":{"type":"integer","format":"int32","description":"Tiempo máximo de incubación recomendado para la receta en minutos","example":720},"refrigerationTime":{"type":"integer","format":"int32","description":"Tiempo de refrigeración recomendado para la receta en minutos","example":60},"difficulty":{"type":"string","description":"Nivel de dificultad de la receta","enum":["BEGINNER","INTERMEDIATE","ADVANCED"],"example":"MEDIO"},"tips":{"type":"string","description":"Consejos para la receta","example":"Asegúrate de mantener una temperatura constante durante todo el proceso"},"ingredients":{"type":"array","description":"Lista de ingredientes para la receta","items":{"$ref":"#/components/schemas/IngredientDto"}}}},"Recipe":{"type":"object","properties":{"id":{"type":"integer","format":"int64","description":"ID de la receta","example":1},"name":{"type":"string","description":"Nombre de la receta","example":"Yogurt natural"},"description":{"type":"string","description":"Descripción de la receta","example":"Receta básica para hacer yogurt natural en casa"},"defaultMilkVolume":{"type":"number","format":"double","description":"Volumen de leche recomendado para la receta","example":1.5},"defaultStarterAmount":{"type":"number","format":"double","description":"Cantidad de cultivo iniciador recomendada para la receta","example":2},"heatingTemperature":{"type":"number","format":"double","description":"Temperatura de calentamiento recomendada para la receta","example":85},"heatingDuration":{"type":"integer","format":"int32","description":"Duración del calentamiento recomendada para la receta","example":30},"inoculationTemperature":{"type":"number","format":"double","description":"Temperatura de inoculación recomendada para la receta","example":45},"incubationTemperature":{"type":"number","format":"double","description":"Temperatura de incubación recomendada para la receta","example":42},"minIncubationTime":{"type":"integer","format":"int32","description":"Tiempo mínimo de incubación recomendado para la receta","example":6},"maxIncubationTime":{"type":"integer","format":"int32","description":"Tiempo máximo de incubación recomendado para la receta","example":12},"refrigerationTime":{"type":"integer","format":"int32","description":"Tiempo de refrigeración recomendado para la receta","example":4},"difficulty":{"type":"string","description":"Nivel de dificultad de la receta","enum":["BEGINNER","INTERMEDIATE","ADVANCED"],"example":"BEGINNER"},"tips":{"type":"string","description":"Consejos adicionales para la receta","example":"Usar leche entera para un yogurt más cremoso"},"active":{"type":"boolean","description":"Indica si la receta está activa","example":true}}},"StartBatchRequest":{"type":"object","properties":{"recipeId":{"type":"integer","format":"int64","description":"ID de la receta a utilizar para el lote de yogurt","example":1},"customMilkVolume":{"type":"number","format":"double","description":"Volumen de leche personalizado para el lote de yogurt (opcional)","example":1.5},"customStartAmount":{"type":"number","format":"double","description":"Cantidad de cultivo iniciador personalizada para el lote de yogurt (opcional)","example":2}}},"TemperatureLog":{"type":"object","properties":{"id":{"type":"integer","format":"int64","description":"ID del registro de temperatura","example":1},"batch":{"$ref":"#/components/schemas/YogurtBatch","description":"Lote de yogurt al que pertenece el registro de temperatura"},"temperature":{"type":"number","format":"double","description":"Temperatura registrada","example":45.0},"recordedAt":{"type":"string","format":"date-time","description":"Fecha y hora en que se registró la temperatura","example":"2024-06-01T14:30:00"},"type":{"type":"string","description":"Tipo de registro de temperatura","enum":["HEATING","COOLING","INCUBATION","REFRIGERATION","MANUAL"],"example":"HEATING"},"notes":{"type":"string","description":"Notas adicionales sobre el registro de temperatura","example":"Temperatura alcanzada durante el proceso de calentamiento"}}},"YogurtBatch":{"type":"object","properties":{"id":{"type":"integer","format":"int64","description":"ID del lote de yogurt","example":1},"batchCode":{"type":"string","description":"Código del lote de yogurt","example":"YB 123456789"},"recipe":{"$ref":"#/components/schemas/Recipe"},"status":{"type":"string","description":"Estado del lote de yogurt","enum":["PREPARING","HEATING","COOLING","INOCULATING","INCUBATION","REFRIGERATION","COMPLETED","FAILED","REFRIGERATING","INCUBATING"],"example":"PREPARING"},"milkVolume":{"type":"number","format":"double","description":"Volumen de leche del lote de yogurt","example":1000.0},"starterAmount":{"type":"number","format":"double","description":"Cantidad de cultivo iniciador del lote de yogurt","example":2.0},"targetTemperature":{"type":"number","format":"double","description":"Temperatura objetivo para el lote de yogurt","example":42.0},"incubationTime":{"type":"integer","format":"int32","description":"Duración de incubación del lote de yogurt en minutos","example":480},"starTime":{"type":"string","format":"date-time","description":"Fecha y hora de inicio del proceso del lote de yogurt","example":"2024-06-01T14:00:00"},"incubarionStarTime":{"type":"string","format":"date-time","description":"Fecha y hora de finalización del proceso del lote de yogurt","example":"2024-06-02T22:00:00"},"incubationEndTime":{"type":"string","format":"date-time","description":"Fecha y hora de finalización del proceso de incubación del lote de yogurt","example":"2024-06-02T22:00:00"},"refrigerationStarTime":{"type":"string","format":"date-time","description":"Fecha y hora de inicio del proceso de refrigeración del lote de yogurt","example":"2024-06-02T22:30:00"},"temperatureLogs":{"type":"array","description":"Lista de registros de temperatura asociados al lote de yogurt","items":{"$ref":"#/components/schemas/TemperatureLog"}},"notes":{"type":"string","description":"Notas adicionales sobre el lote de yogurt","example":"Este lote es para una promoción especial"},"createdAt":{"type":"string","format":"date-time","description":"Indica si el lote de yogurt es un lote de prueba","example":"false"},"updateAt":{"type":"string","format":"date-time","description":"Fecha y hora de última actualización del lote de yogurt","example":"2024-06-01T15:00:00"}}},"FailRequest":{"type":"object","properties":{"reason":{"type":"string","description":"Razón por la cual el lote de yogurt ha fallado","example":"Contaminación detectada durante la incubación"}}},"Dashboard":{"type":"object","properties":{"batchCounts":{"type":"object","additionalProperties":{"type":"integer","format":"int64"},"description":"Conteo de lotes de yogurt por estado (activo, completado, fallido)","example":{"ACTIVE":3,"COMPLETED":5,"FAILED":1}},"activeBatchCount":{"type":"integer","format":"int64","description":"Número de lotes de yogurt actualmente activos","example":3},"completedToday":{"type":"integer","format":"int32","description":"Número de lotes de yogurt completados hoy","example":2}}},"TemperatureSummary":{"type":"object","properties":{"currentTemperature":{"type":"number","format":"double","description":"Temperatura actual del lote de yogurt","example":42.5},"maximumTemperature":{"type":"number","format":"double","description":"Temperatura máxima registrada durante la incubación","example":43.0},"minimumTemperature":{"type":"number","format":"double","description":"Temperatura mínima registrada durante la incubación","example":41.5},"averageTemperature":{"type":"number","format":"double","description":"Temperatura promedio registrada durante la incubación","example":42.0}}}}}}